requirements.txt
autopep8==1.5.7
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
pycodestyle==2.7.0
toml==0.10.2
